+++ src/SFA.DAS.EmploymentCheck.Api/Mediators/Commands/RegisterCheckCommand/RegisterCheckCommand.cs
@@ -8,7 +8,7 @@ namespace SFA.DAS.EmploymentCheck.Api.Mediators.Commands.RegisterCheckCommand
         public Guid CorrelationId { get; set; }
         public string CheckType { get; set; }
         public long Uln { get; set; }
-        public int ApprenticeshipAccountId { get; set; }
+        public long ApprenticeshipAccountId { get; set; }
         public long? ApprenticeshipId { get; set; }
         public DateTime MinDate { get; set; }
         public DateTime MaxDate { get; set; }
